feature: move analytics to middleware
